check if VSCode is installed

+++ scripts/osx.sh
@@ -2,51 +2,55 @@
 # OSX.sh
 
 # -- Screenshots ------------------------------------------------------------------
-  screenshotsFolder="$HOME/Documents/screenshots"
-  if get_boolean_response "Create a folder for the screenshots at $screenshotsFolder?"; then
-   mkdir "$screenshotsFolder"
-   defaults write com.apple.screencapture location  "$screenshotsFolder"
-  else
+screenshotsFolder="$HOME/Documents/screenshots"
+if get_boolean_response "Create a folder for the screenshots at $screenshotsFolder?"; then
+    mkdir "$screenshotsFolder"
+    defaults write com.apple.screencapture location  "$screenshotsFolder"
+else
     echo_item "Skip changing default screenshots folder" "red"
-  fi
-  
+fi
+
 # -- Homebrew ------------------------------------------------------------------
 
 if exists "brew"; then
-  echo_item "Homebrew is already installed" green
+    echo_item "Homebrew is already installed" green
 else
-  if get_boolean_response "Do you want to install Homebrew?"; then
-    ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-  else
-    echo_item "Skipping Homebrew install" "red"
-  fi
+    if get_boolean_response "Do you want to install Homebrew?"; then
+        ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+    else
+        echo_item "Skipping Homebrew install" "red"
+    fi
 fi
 
 echo ""
 
 if exists "brew"; then
-# -- Fonts ------------------------------------------------------------------
-  if get_boolean_response "Do you want to initialize cask fonts ?"; then
-    brew tap homebrew/cask-fonts                  # you only have to do this once!
-  else
-    echo_item "Skip font installation" "red"
-  fi
-  echo ""
-  if get_boolean_response "Do you want to install some fonts?"; then
-      brew cask install font-inconsolata-nerd-font
-      brew cask install font-inconsolata
-      brew cask install font-fira-code
-  else
-    echo_item "Skipping fonts install" red
-  fi
-# -- Software (desktop apps) ---------------------------------------------------------------------
-  if get_boolean_response "Do you want to install Visual Studio code?"; then
-    brew cask install visual-studio-code
-  else
-    echo_item "Skipping visual studio code install" red
-  fi
+    # -- Fonts ------------------------------------------------------------------
+    if get_boolean_response "Do you want to initialize cask fonts ?"; then
+        brew tap homebrew/cask-fonts                  # you only have to do this once!
+    else
+        echo_item "Skip font installation" "red"
+    fi
+    echo ""
+    if get_boolean_response "Do you want to install some fonts?"; then
+        brew cask install font-inconsolata-nerd-font
+        brew cask install font-inconsolata
+        brew cask install font-fira-code
+    else
+        echo_item "Skipping fonts install" red
+    fi
+    # -- Software (desktop apps) ---------------------------------------------------------------------
+    if exists "code"; then
+        echo_item "Visual studio Code is already installed" green
+    else
+        if get_boolean_response "Do you want to install Visual Studio code?"; then
+            brew cask install visual-studio-code
+        else
+            echo_item "Skipping visual studio code install" red
+        fi
+    fi
 else
-  echo_item "Brew is not installed, skipping fonts and desktop apps installation" red
+    echo_item "Brew is not installed, skipping fonts and desktop apps installation" red
 fi
 
 echo ""
@@ -54,50 +58,50 @@ echo ""
 # -- Node/NPM/Yarn -----------------------------------------------------------------------
 
 if exists "npm"; then
-  echo_item "npm is already installed" green
+    echo_item "npm is already installed" green
 else
-  if get_boolean_response "Do you want to install node+npm?"; then
-    brew install node
-    source ./node.sh
-  else
-    echo_item "Skipping npm install" red
-  fi
+    if get_boolean_response "Do you want to install node+npm?"; then
+        brew install node
+        source ./node.sh
+    else
+        echo_item "Skipping npm install" red
+    fi
 fi
 
 if exists "yarn"; then
-  echo_item "yarn is already installed" green
+    echo_item "yarn is already installed" green
 else
-  if get_boolean_response "Do you want to install yarn?"; then
-    brew install yarn --without-node
-  else
-    echo_item "Skipping yarn install" red
-  fi
+    if get_boolean_response "Do you want to install yarn?"; then
+        brew install yarn --without-node
+    else
+        echo_item "Skipping yarn install" red
+    fi
 fi
 
 echo ""
 # -- rlwrap -----------------------------------------------------------------------
 
 if exists "rlwrap"; then
-  echo_item "rlwrap is already installed" green
+    echo_item "rlwrap is already installed" green
 else
-  if get_boolean_response "Do you want to install rlwrap?"; then
-    brew install rlwrap
-  else
-    echo_item "Skipping rlwrap install" red
-  fi
+    if get_boolean_response "Do you want to install rlwrap?"; then
+        brew install rlwrap
+    else
+        echo_item "Skipping rlwrap install" red
+    fi
 fi
 
 echo ""
 # -- telnet -----------------------------------------------------------------------
 
 if exists "telnet"; then
-  echo_item "telnet is already installed" green
+    echo_item "telnet is already installed" green
 else
-  if get_boolean_response "Do you want to install telnet?"; then
-    brew install telnet
-  else
-    echo_item "Skipping telnet install" red
-  fi
+    if get_boolean_response "Do you want to install telnet?"; then
+        brew install telnet
+    else
+        echo_item "Skipping telnet install" red
+    fi
 fi
 
 echo ""
@@ -109,13 +113,13 @@ echo ""
 # -- zsh -----------------------------------------------------------------------
 
 if exists "zsh"; then
-  echo_item "zsh is already installed" green
+    echo_item "zsh is already installed" green
 else
-  if get_boolean_response "Do you want to install zsh?"; then
-    brew install zsh
-  else
-    echo_item "Skipping zsh install" red
-  fi
+    if get_boolean_response "Do you want to install zsh?"; then
+        brew install zsh
+    else
+        echo_item "Skipping zsh install" red
+    fi
 fi
 
 echo ""
@@ -123,22 +127,22 @@ echo ""
 # -- Neovim --------------------------------------------------------------------
 
 if exists "nvim"; then
-  echo_item "Neovim is already installed" green
+    echo_item "Neovim is already installed" green
 else
-  if get_boolean_response "Do you want to install Neovim?"; then
-    brew install --HEAD neovim
-    curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-  else
-    echo_item "Skipping Neovim install" red
-  fi
+    if get_boolean_response "Do you want to install Neovim?"; then
+        brew install --HEAD neovim
+        curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+    else
+        echo_item "Skipping Neovim install" red
+    fi
 fi
 
 # -- ssh --------------------------------------------------------------------
 if test -f ~/.ssh/id_rsa; then
-  if get_boolean_response "Seems there is an existing id_rsa key, do you want to import it ?"; then
-    eval "$(ssh-agent -s)"
-    ssh-add -K ~/.ssh/id_rsa
-  else
-    echo_item "Skipping ssh keys import" red
-  fi
+    if get_boolean_response "Seems there is an existing id_rsa key, do you want to import it ?"; then
+        eval "$(ssh-agent -s)"
+        ssh-add -K ~/.ssh/id_rsa
+    else
+        echo_item "Skipping ssh keys import" red
+    fi
 fi
